fix some bugs and add tests for rip relative modrm

+++ src/lib/arch/x86/ctx.c
@@ -210,7 +210,7 @@ err_t pis_x86_lift(
     const pis_x86_ctx_t* ctx,
     const u8* machine_code,
     size_t machine_code_len,
-    uintptr_t machine_code_addr,
+    u64 machine_code_addr,
     pis_lift_result_t* result
 ) {
     err_t err = SUCCESS;
